+++ src/components/shared/basic/SearchResult/index.tsx
@@ -38,7 +38,7 @@ export const SearchResult = ({
   )
   return (
     <Box sx={{ display: 'flex' }}>
-      <Paper elevation={5} sx={{ width: 700, padding: 4, borderRadius: 4 }}>
+      <Box className="search-group" sx={{ borderRadius: 4 }}>
         {groupList.map((item: [string, SearchItem[]], index: number) => (
           <SearchResultGroup
             key={item[0]}
@@ -48,7 +48,7 @@ export const SearchResult = ({
             isFirst={index === 0}
           />
         ))}
-      </Paper>
+      </Box>
     </Box>
   )
 }